Moved most files back to original folder to simplify integrating new changes from master (will redo later)
