+++ vessel/utils/checksum.py
@@ -33,7 +33,7 @@ from typing import Any, Optional
 
 import magic
 
-from vessel.diff.helpers.diffoscope import build_diff_lookup
+from vessel.utils.diffoscope import build_diff_lookup
 
 logger = getLogger(__name__)
